requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.24.0
xformers>=0.0.22
Pillow>=10.0.0
flask>=3.0.0
